+++ Model/Ui/GenericConfigProvider.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 
 namespace MultiSafepay\ConnectCore\Model\Ui;
 
+use Exception;
 use Magento\Checkout\Model\ConfigProviderInterface;
 use Magento\Checkout\Model\Session;
 use Magento\Framework\App\Config\Storage\WriterInterface;
@@ -322,13 +323,31 @@ class GenericConfigProvider implements ConfigProviderInterface
     }
 
     /**
+     * Retrieve issuers through API request if payment method is turned on
+     *
      * @return array
      * @throws ClientExceptionInterface
+     * @throws Exception
      */
     public function getIssuers(): array
     {
         $issuers = [];
-
+        
+        $paymentConfig = $this->getPaymentConfig($this->getStoreIdFromCheckoutSession());
+        
+        if (!$paymentConfig) {
+            $this->logger->debug('Payment config not found when retrieving issuers');
+        }
+        
+        if (!isset($paymentConfig['active'])) {
+            $this->logger->debug('Could not check if payment method is activated when retrieving issuers');
+        }
+        
+        if (!$paymentConfig['active']) {
+            // Payment method not activated, issuer request not sent
+            return [];
+        }
+        
         if ($multiSafepaySdk = $this->getSdk()) {
             try {
                 $issuerListing = $multiSafepaySdk->getIssuerManager()->getIssuersByGatewayCode($this->getGatewayCode());